# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Diploma RAG API service documentation.",
        "title": "Diploma API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/larek-tech/diploma/blob/api/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/source/permissions/roles/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns list of roles permitted to source.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "source"
                ],
                "summary": "Get permitted roles.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Requested source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permitted roles",
                        "schema": {
                            "$ref": "#/definitions/pb.Source"
                        }
                    },
                    "404": {
                        "description": "Source not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates list of roles permitted to source.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "source"
                ],
                "summary": "Update permitted roles.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Requested source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New list of permitted roles",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.PermittedRoles"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated roles permissions",
                        "schema": {
                            "$ref": "#/definitions/pb.PermittedRoles"
                        }
                    },
                    "404": {
                        "description": "Source not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/source/permissions/users/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns list of users permitted to source.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "source"
                ],
                "summary": "Get permitted users.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Requested source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permitted users",
                        "schema": {
                            "$ref": "#/definitions/pb.Source"
                        }
                    },
                    "404": {
                        "description": "Source not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates list of users permitted to source.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "source"
                ],
                "summary": "Update permitted users.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Requested source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New list of permitted users",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.PermittedUsers"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated users permissions",
                        "schema": {
                            "$ref": "#/definitions/pb.PermittedUsers"
                        }
                    },
                    "404": {
                        "description": "Source not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/source/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns information about source and its update parameters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "source"
                ],
                "summary": "Get source.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Requested source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Source",
                        "schema": {
                            "$ref": "#/definitions/pb.Source"
                        }
                    },
                    "400": {
                        "description": "Failed to get source",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Source not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update source information or params for update job.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "source"
                ],
                "summary": "Update source.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update params",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.UpdateSourceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Source updated",
                        "schema": {
                            "$ref": "#/definitions/pb.Source"
                        }
                    },
                    "400": {
                        "description": "Failed to update source",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Source not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete domain source by ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "source"
                ],
                "summary": "Delete source.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Source deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Failed to delete source",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Source not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pb.CronFormat": {
            "type": "object",
            "properties": {
                "dayOfMonth": {
                    "type": "integer"
                },
                "dayOfWeek": {
                    "type": "integer"
                },
                "hour": {
                    "type": "integer"
                },
                "minute": {
                    "type": "integer"
                },
                "month": {
                    "type": "integer"
                }
            }
        },
...
        "pb.PermittedRoles": {
            "type": "object",
            "properties": {
                "resourceId": {
                    "type": "integer"
                },
                "roleIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "pb.PermittedUsers": {
            "type": "object",
            "properties": {
                "resourceId": {
                    "type": "integer"
                },
                "userIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "pb.Source": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "createdAt": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                },
                "credentials": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/pb.SourceStatus"
                },
                "title": {
                    "type": "string"
                },
                "typ": {
                    "$ref": "#/definitions/pb.SourceType"
                },
                "updateParams": {
                    "$ref": "#/definitions/pb.UpdateParams"
                },
                "updatedAt": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "pb.SourceStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "SourceStatus_STATUS_UNDEFINED",
                "SourceStatus_STATUS_READY",
                "SourceStatus_STATUS_PARSING",
                "SourceStatus_STATUS_FAILED"
            ]
        },
        "pb.SourceType": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4
            ],
            "x-enum-varnames": [
                "SourceType_TYPE_UNDEFINED",
                "SourceType_TYPE_WEB",
                "SourceType_TYPE_SINGLE_FILE",
                "SourceType_TYPE_ARCHIVED_FILES",
                "SourceType_TYPE_WITH_CREDENTIALS"
            ]
        },
...
        "pb.UpdateParams": {
            "type": "object",
            "properties": {
                "cron": {
                    "$ref": "#/definitions/pb.CronFormat"
                },
                "everyPeriod": {
                    "type": "integer"
                }
            }
        },
...
        "pb.UpdateSourceRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "credentials": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "sourceId": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updateParams": {
                    "$ref": "#/definitions/pb.UpdateParams"
                }
            }
        },
...
        "timestamppb.Timestamp": {
            "type": "object",
            "properties": {
                "nanos": {
                    "description": "Non-negative fractions of a second at nanosecond resolution. Negative\nsecond values with fractions must still have non-negative nanos values\nthat count forward in time. Must be from 0 to 999,999,999\ninclusive.",
                    "type": "integer"
                },
                "seconds": {
                    "description": "Represents seconds of UTC time since Unix epoch\n1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to\n9999-12-31T23:59:59Z inclusive.",
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}